PR bug fixes
