+++ .gitignore
@@ -37,3 +37,24 @@ Zc*.mat
 
 ChatGPT*.html
 ChatGPT*.txt
+
+# OS-generated files
+.DS_Store
+
+# Compiled binaries and object files
+*.o
+*.a
+*.so
+*.dll
+*.exe
+*.out
+
+# Build output directories (example, adjust as needed)
+/bin/
+/build/
+/dist/
+
+# Other common temporary files
+*.log
+*.tmp
+*.npy
