+++ pyCoilGen/sub_functions/gradient_former.py
@@ -1,0 +1,381 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.patches import Rectangle, Circle, FancyArrow
+import trimesh
+from scipy.interpolate import splprep, splev
+from pyCoilGen.plotting.plot_wire_loops import plot_gerber_paths, plot_stl_patch
+
+# ==========================================================
+# 3D printer G-code generation for gradient coil holder
+# ==========================================================
+
+def generate_gradient_former(
+        coil_parts,
+        output_prefix="gradient_plate",
+        plate_diameter_m=0.1524,
+        plate_thickness_m=0.003,
+        wire_width_m=0.002,
+        display_debug_plots=True,
+        save_loop_coords=True):
+
+    """
+    Export gradient coil loops as RS-274X Gerber for CNC fabrication
+    and provide 2D visualization including wire width for overlap checks.
+
+    Also exports complement Gerber files where the copper region is the
+    inverse of the wire paths for isolation milling.
+    """
+
+
+
+    mm = 1000.0
+    plate_radius_mm = plate_diameter_m*mm/2
+    wire_width_mm = wire_width_m*mm
+
+    #-------------------------
+    # Generate the plate STL for merging later if needed
+    #-------------------------
+
+
+
+    # ----------------------------------------
+    # Helper to ensure minimum spacing between points
+    # ----------------------------------------
+    def enforce_global_spacing(points, min_dist_mm=0.1):
+        filtered=[points[0]]
+        for p in points[1:]:
+            if np.linalg.norm(p-filtered[-1])>min_dist_mm:
+                filtered.append(p)
+        return np.array(filtered)
+
+    # ----------------------------------------
+    # Helper to generate circular plate
+    # ----------------------------------------
+    def generate_plate_circle(radius_mm, npts=720):
+        th = np.linspace(0,2*np.pi,npts)
+        x = radius_mm*np.cos(th)
+        y = radius_mm*np.sin(th)
+        return np.column_stack([x,y])
+
+    # ----------------------------------------
+    # Process each coil plate
+    # ----------------------------------------
+    for part_index, part in enumerate(coil_parts):
+
+        print(f"\nProcessing plate {part_index}")
+
+        gerber_paths = []
+
+        # ----------------------------------------
+        # Collect paths
+        # ----------------------------------------
+        for loop_index, loop in enumerate(part.wire_loops):
+
+            pts_mm = np.asarray(loop["path"])[:, :2]*mm
+            pts_mm = enforce_global_spacing(pts_mm, min_dist_mm=0.5)
+
+            gerber_paths.append((pts_mm, loop["sign"]))
+
+            if save_loop_coords:
+                coords = np.column_stack([
+                    pts_mm[:,0],
+                    pts_mm[:,1],
+                    np.zeros(len(pts_mm))
+                ])
+
+                fname = f"{output_prefix}_plate{part_index}_loop{loop_index}.txt"
+                np.savetxt(fname, coords, fmt="%.2f")
+
+
+
+        # ----------------------------------------
+        # Write standard Gerber (tracks)
+        # ----------------------------------------
+        gerber_fname = f"{output_prefix}_plate{part_index}.gbr"
+
+        with open(gerber_fname, "w") as f:
+
+            f.write("G04 Gradient coil generated by Python*\n")
+            f.write("%MOMM*%\n")
+            f.write("%FSLAX24Y24*%\n")
+
+            f.write(f"%ADD10C,{wire_width_mm:.4f}*%\n")
+
+            f.write("D10*\n")
+
+            for path, sign in gerber_paths:
+
+                x0, y0 = path[0]
+
+                f.write(f"X{int(x0*10000):07d}Y{int(y0*10000):07d}D02*\n")
+
+                for p in path[1:]:
+                    x, y = p
+                    f.write(f"X{int(x*10000):07d}Y{int(y*10000):07d}D01*\n")
+
+            f.write("M02*\n")
+
+        print("Saved Gerber:", gerber_fname)
+
+        # ----------------------------------------
+        # Write complement Gerber (plate minus tracks)
+        # ----------------------------------------
+        complement_fname = f"{output_prefix}_plate{part_index}_complement.gbr"
+
+        plate_circle = generate_plate_circle(plate_radius_mm)
+
+        with open(complement_fname, "w") as f:
+
+            f.write("G04 Complement copper layer*\n")
+            f.write("%MOMM*%\n")
+            f.write("%FSLAX24Y24*%\n")
+
+            # aperture for tracks
+            f.write(f"%ADD10C,{wire_width_mm:.4f}*%\n")
+
+            # ---- Filled plate region ----
+            f.write("%LPD*%\n")
+            f.write("G36*\n")
+
+            x0, y0 = plate_circle[0]
+
+            f.write(f"X{int(x0*10000):07d}Y{int(y0*10000):07d}D02*\n")
+
+            for p in plate_circle[1:]:
+                x,y = p
+                f.write(f"X{int(x*10000):07d}Y{int(y*10000):07d}D01*\n")
+
+            f.write(f"X{int(x0*10000):07d}Y{int(y0*10000):07d}D01*\n")
+
+            f.write("G37*\n")
+
+            # ---- Subtract wire paths ----
+            f.write("%LPC*%\n")
+            f.write("D10*\n")
+
+            for path,_ in gerber_paths:
+
+                x0,y0 = path[0]
+
+                f.write(f"X{int(x0*10000):07d}Y{int(y0*10000):07d}D02*\n")
+
+                for p in path[1:]:
+                    x,y = p
+                    f.write(f"X{int(x*10000):07d}Y{int(y*10000):07d}D01*\n")
+
+            f.write("M02*\n")
+
+        print("Saved Complement Gerber:", complement_fname)
+
+        # ----------------------------------------
+        # Visualization
+        # ----------------------------------------
+        if display_debug_plots:
+            plot_gerber_paths(gerber_paths, plate_radius_mm, wire_width_mm, part_index)
+        # ----------------------------------------
+        # STL GENERATION (FINAL: all features)
+        # ----------------------------------------
+        try:
+
+
+            def smooth_path(x, y, smoothing=5.0, n_points=300):
+                tck, _ = splprep([x, y], s=smoothing, per=True)
+                u = np.linspace(0, 1, n_points)
+                x_new, y_new = splev(u, tck)
+                return np.array(x_new), np.array(y_new)
+
+            all_segments = []
+            hole_meshes = []
+            strain_relief_segments = []
+            numbering_meshes = []
+
+            plate_height_mm = plate_thickness_m * mm
+            hole_radius = wire_width_mm * 0.8
+            hole_height = plate_height_mm * 2.0
+            offset_dist = wire_width_mm * 1.5
+            strain_length = wire_width_mm * 4.0
+            tube_radius = wire_width_mm / 2.0
+
+            # ---- Plate ----
+            plate = trimesh.creation.cylinder(
+                radius=plate_radius_mm,
+                height=plate_height_mm,
+                sections=128
+            )
+            plate.apply_translation([0, 0, plate_height_mm / 2])
+
+            # ---- Alignment pins (snap-fit) ----
+            pin_radius = wire_width_mm
+            pin_height = plate_height_mm * 1.2
+            pin_offset = plate_radius_mm * 0.85
+            pin_angles = [0, np.pi/2, np.pi, 3*np.pi/2]  # 4 pins
+
+            pin_meshes = []
+            for a in pin_angles:
+                x = pin_offset * np.cos(a)
+                y = pin_offset * np.sin(a)
+                pin = trimesh.creation.cylinder(
+                    radius=pin_radius,
+                    height=pin_height
+                )
+                pin.apply_translation([x, y, plate_height_mm / 2])
+                pin_meshes.append(pin)
+
+            # ---- Process each loop ----
+            for loop_index, (path, sign) in enumerate(gerber_paths):
+
+                x_raw, y_raw = path[:,0], path[:,1]
+                points = smooth_path(x_raw, y_raw, smoothing=5.0)
+
+                points_3d = np.column_stack([points[0], points[1], np.zeros_like(points[0])])
+
+                # ---- Tube segments ----
+                for i in range(len(points_3d)-1):
+                    p1, p2 = points_3d[i], points_3d[i+1]
+                    if np.linalg.norm(p2 - p1) < 0.5:
+                        continue
+                    seg = trimesh.creation.cylinder(
+                        radius=tube_radius,
+                        segment=[p1, p2]
+                    )
+                    # optional multi-layer: raise every other loop slightly
+                    if loop_index % 2 == 1:
+                        seg.apply_translation([0,0,0.2])
+                    all_segments.append(seg)
+
+                # ---- Terminal holes ----
+                def tangent(p0, p1):
+                    v = p1 - p0
+                    n = np.linalg.norm(v)
+                    return v/n if n>0 else np.array([1.0,0.0])
+                def perp(v): return np.array([-v[1], v[0]])
+
+                t_start = tangent(path[0], path[1])
+                t_end = tangent(path[-2], path[-1])
+                start_pt = path[0] + perp(t_start)*offset_dist
+                end_pt = path[-1] + perp(t_end)*offset_dist
+
+                for pt in [start_pt, end_pt]:
+                    hole = trimesh.creation.cylinder(
+                        radius=hole_radius,
+                        height=hole_height
+                    )
+                    hole.apply_translation([pt[0], pt[1], plate_height_mm / 2])
+                    hole_meshes.append(hole)
+
+                # ---- Strain relief ----
+                for base_pt, t_vec in [(start_pt, t_start), (end_pt, t_end)]:
+                    p1 = np.array([base_pt[0], base_pt[1], 0])
+                    p2 = p1 + np.append(t_vec * strain_length, 0)
+                    relief = trimesh.creation.cylinder(
+                        radius=tube_radius,
+                        segment=[p1, p2]
+                    )
+                    strain_relief_segments.append(relief)
+
+                # ---- Loop numbering ----
+                # try:
+                #     num_text = trimesh.creation.text(
+                #         text=str(loop_index+1),
+                #         height=wire_width_mm*2,
+                #         depth=0.5
+                #     )
+                #     # place number at first point of loop, slightly above plate
+                #     num_text.apply_translation([
+                #         path[0][0],
+                #         path[0][1],
+                #         plate_height_mm - 0.3
+                #     ])
+                #     numbering_meshes.append(num_text)
+                # except Exception as e:
+                #     print(f"⚠️ Loop numbering failed for loop {loop_index+1}: {e}")
+
+            # ---- Combine all cutters ----
+            tube = trimesh.util.concatenate(all_segments)
+            if strain_relief_segments:
+                relief_mesh = trimesh.util.concatenate(strain_relief_segments)
+                tube = trimesh.util.concatenate([tube, relief_mesh])
+            cutter_meshes = [tube] + hole_meshes + numbering_meshes + pin_meshes
+            cutter = trimesh.util.concatenate(cutter_meshes)
+
+            # ---- Offset tube slightly into plate for engraving/groove ----
+            cutter.apply_translation([0,0,plate_height_mm*0.75])
+
+            # # ---- Plate engraving ----
+            # label = "TOP" if part_index==0 else "BOTTOM"
+            # try:
+            #     text_mesh = trimesh.creation.text(
+            #         text=label,
+            #         height=plate_radius_mm*0.15,
+            #         depth=0.5
+            #     )
+            #     text_mesh.apply_translation([
+            #         -plate_radius_mm*0.3,
+            #         -plate_radius_mm*0.8,
+            #         plate_height_mm - 0.5
+            #     ])
+            #     cutter = trimesh.util.concatenate([cutter, text_mesh])
+            # except:
+            #     print("⚠️ Plate engraving skipped")
+
+            # ---- Boolean difference ----
+            result = plate.difference(cutter)
+            if result is None or len(result.vertices)==0:
+                print("❌ Boolean failed")
+                continue
+
+            # ---- Export STL ----
+            stl_fname = f"{output_prefix}_plate{part_index}_final_print.stl"
+            result.export(stl_fname)
+            print(f"✅ Final STL saved: {stl_fname}")
+
+
+            # ----------------------------------------
+            # DEBUG VISUALIZATION (Matplotlib + terminals + strain relief + loop numbers)
+            # ----------------------------------------
+            if display_debug_plots:
+                plot_stl_patch(gerber_paths, plate_radius_mm, wire_width_mm, part_index)
+               
+        except Exception as e:
+            print("❌ STL generation failed:", e)
+
+    print("\nGradient former coordinate export complete with Gerber and visualization.")
+
+
+
+
+# ===============================
+# CHECK SPHERICAL HARMONICS
+#================================
+
+
+def run_spherical_harmonic_diagnostic(result):
+
+    coords = result.target_field.coords
+    field = result.coil_gradient.gradient_in_target_direction
+
+    x,y,z = coords
+
+    r = np.sqrt(x**2+y**2+z**2)
+    theta = np.arccos(z/r)
+    phi = np.arctan2(y,x)
+
+    max_l = 3
+
+    coeffs = {}
+
+    for l in range(max_l+1):
+        for m in range(-l,l+1):
+
+            Ylm = sph_harm_y(l,m,phi,theta).real
+            a = np.sum(field*Ylm)/np.sum(Ylm**2)
+
+            coeffs[(l,m)] = a
+
+    print("\n===== Spherical Harmonic Content =====")
+
+    for k,v in coeffs.items():
+        print(f"l={k[0]}, m={k[1]} : {v:.4e}")
+
+
+
